+++ app/controllers/application_controller.rb
@@ -17,6 +17,7 @@ class ApplicationController < ActionController::Base
 
   include ControllerAuthorization
   include ActivityLogger
+  include LogRagePayloadBehavior
   include Pagy::Backend
   protect_from_forgery with: :exception
 
@@ -108,16 +109,7 @@ class ApplicationController < ActionController::Base
 
   def append_info_to_payload(payload)
     super
-    payload[:server_protocol] = request.env['SERVER_PROTOCOL']
-    payload[:remote_ip] = request.remote_ip
-    payload[:ip] = request.ip
-    payload[:remote_addr] = request.env['REMOTE_ADDR']
-    payload[:x_forwarded_for] = request.headers['HTTP_X_FORWARDED_FOR']
-    payload[:session_id] = request.env['rack.session.record'].try(:session_id)
     payload[:user_id] = current_user&.id
-    payload[:pid] = Process.pid
-    payload[:request_id] = request.uuid
-    payload[:request_start] = request.headers['HTTP_X_REQUEST_START'].try(:gsub, /\At=/, '')
   end
 
   def info_for_paper_trail
